Correct comments on the moonlight program =========================================
In late 2019 I carried out some tests on the moonlight program
which showed that the errors in the positions of the Moon
and Sun were less than 0.5 degrees - and that this level
of error could be expected at least until 2100.

My original note was unclear about the accuracy to expect so the
new version gives confidence that the program can continue to
be used.